Moved slices and other unions
