+++ libs/api/domains/cms/src/lib/models/adgerdirFrontpage.model.ts
@@ -1,18 +1,11 @@
 /* eslint-disable @typescript-eslint/no-inferrable-types */
 import { Field, ObjectType, ID } from '@nestjs/graphql'
-import { ApolloError } from 'apollo-server-express'
 import {
   IVidspyrnaFrontpage,
-  IVidspyrnaFeaturedNews,
-  IVidspyrnaFlokkur,
 } from '../generated/contentfulTypes'
-
-import { mapDocument } from './slice.model'
-import { mapAdgerdirFeaturedNewsSlice } from './adgerdirFeaturedNewsSlice.model'
-import { mapAdgerdirGroupSlice } from './adgerdirGroupSlice.model'
 import { Image, mapImage } from './image.model'
-import { SliceUnion } from '../unions/slice.union'
-import { AdgerdirSliceUnion } from '../unions/adgerdirSlice.union'
+import { mapDocument, SliceUnion } from '../unions/slice.union'
+import { AdgerdirSliceUnion, mapAdgerdirSliceUnion } from '../unions/adgerdirSlice.union'
 import { SystemMetadata } from '@island.is/shared/types'
 
 @ObjectType()
@@ -39,27 +32,6 @@ export class AdgerdirFrontpage {
   featuredImage?: Image
 }
 
-type AdgerdirSliceTypes = IVidspyrnaFeaturedNews | IVidspyrnaFlokkur
-
-export const mapAdgerdirSlice = (
-  slice: AdgerdirSliceTypes,
-): typeof AdgerdirSliceUnion => {
-  const id = slice?.sys?.contentType?.sys?.id ?? ''
-
-  switch (id) {
-    case 'vidspyrnaFeaturedNews':
-      return mapAdgerdirFeaturedNewsSlice(slice as IVidspyrnaFeaturedNews)
-
-    case 'vidspyrnaFlokkur':
-      return mapAdgerdirGroupSlice(slice as IVidspyrnaFlokkur)
-
-    default:
-      throw new ApolloError(
-        `Can not convert to slice in mapAdgerdirFrontpage -> mapAdgerdirSlice`,
-      )
-  }
-}
-
 export const mapAdgerdirFrontpage = ({
   sys,
   fields,
@@ -76,7 +48,7 @@ export const mapAdgerdirFrontpage = ({
       : [],
   slices: fields?.slices
     ? fields.slices
-        .filter((x) => x.sys?.contentType?.sys?.id)
-        .map(mapAdgerdirSlice)
+      .filter((x) => x.sys?.contentType?.sys?.id)
+      .map(mapAdgerdirSliceUnion)
     : [],
 })
